[EBC] [WIP] Fix function call arguments passing.
This commit fixes arguments passing in function calls.
The EBC calling convention is CDECL, so all arguments are passed by stack.
The current implementation uses stores to push the arguments,
instead of using PUSH{32,64,n}.
The implementation can not manipulate the r0 stack pointer correctly.
It must setup the r0 before CALL instruction. (e.g. add64 r0, r0 -8)
In addition, the order of pushed arguments are incorrect,
it must passed in reverse order according to CDECL.
So, it is still incomplete work.

Signed-off-by: Akira Moroo <retrage01@gmail.com>